Newsrooms_NLP/code/requirements.txt
os
sys
python-docx